Extract `utils.h`
