+++ ext/saturn/graph.c
@@ -1,5 +1,6 @@
 #include "graph.h"
 #include "ruby/internal/value_type.h"
+#include "utils.h"
 
 static VALUE cGraph;
 
@@ -19,32 +20,6 @@ static const rb_data_type_t graph_type = {
 static VALUE rb_graph_alloc(VALUE klass) {
     void *graph = sat_graph_new();
     return TypedData_Wrap_Struct(klass, &graph_type, graph);
-}
-
-// Convert a Ruby array of strings into a double char pointer so that we can
-// pass that to Rust. This copies the data so it must be freed
-static char **str_array_to_char(VALUE array, size_t length) {
-    char **converted_array = malloc(length * sizeof(char *));
-
-    for (size_t i = 0; i < length; i++) {
-        VALUE item = rb_ary_entry(array, i);
-        const char *string = StringValueCStr(item);
-
-        converted_array[i] = malloc(strlen(string) + 1);
-        strcpy(converted_array[i], string);
-    }
-
-    return converted_array;
-}
-
-// Verify that the Ruby object is an array of strings or raise `TypeError`
-static void check_array_of_strings(VALUE array) {
-    Check_Type(array, T_ARRAY);
-
-    for (long i = 0; i < RARRAY_LEN(array); i++) {
-        VALUE item = rb_ary_entry(array, i);
-        Check_Type(item, T_STRING);
-    }
 }
 
 // Graph#index_all: (Array[String] file_paths) -> String?
